Commited all forgotten files.

+++ lentaruAndTests/src/cz/fit/lentaruand/test/NewsDaoTest.java
@@ -8,8 +8,8 @@ import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 
+import android.content.ContentResolver;
 import android.content.Context;
-import android.database.sqlite.SQLiteDatabase;
 import android.test.AndroidTestCase;
 import android.test.suitebuilder.annotation.SmallTest;
 import cz.fit.lentaruand.data.Link;
@@ -20,66 +20,65 @@ import cz.fit.lentaruand.data.dao.NewsLinksDao;
 import cz.fit.lentaruand.data.db.LentaDbHelper;
 
 public class NewsDaoTest extends AndroidTestCase {
-
-	private LentaDbHelper dbHelper;
-	private Context context;
 	private News testNews;
 	private Date date;
+	private static boolean initialized;
 	
-	@Override
-	protected void setUp() throws Exception {
-		super.setUp();
-		context = getContext();
-		context.deleteDatabase(LentaDbHelper.DATABASE_NAME);
+	public NewsDaoTest() {
 		date = new Date();
 		Link l1 = new Link("http://www.sky.com", "Link to the heaven", date);
 		Link l2 = new Link("http://www.sky1.com", "Link to the heaven 1", date);
 		Link l3 = new Link("http://www.sky2.com", "Link to the heaven 2", date);
 		testNews = new News("guid1", "News 1", "http://www.1.ru", "Brief news info", "Full news text", 
 				date, "http://www.image.com/image.png", "Image caption", "Photo: PK", Arrays.asList(l1, l2, l3), Rubrics.CULTURE, true);
-		dbHelper = new LentaDbHelper(context);
+	}
+
+	@Override
+	protected void setUp() throws Exception {
+		if (!initialized) {
+			Context context = getContext();
+			context.deleteDatabase(LentaDbHelper.DATABASE_NAME);
+			initialized = true;
+		}
+		
+		super.setUp();
+	}
+
+	@Override
+	protected void tearDown() throws Exception {
+		super.tearDown();
+		//context = getContext();
+		//context.deleteDatabase(LentaDbHelper.DATABASE_NAME);
 	}
 
 	private long createNews(String guid) {
-		SQLiteDatabase db = dbHelper.getWritableDatabase();
-		try {
-			NewsDao newsDao = new NewsDao();
-			testNews.setGuid(guid);
-			return newsDao.create(db, testNews);
-		} finally {
-			db.close();
-		}
+		ContentResolver cr = getContext().getContentResolver();
+		NewsDao newsDao = new NewsDao(cr);
+		testNews.setGuid(guid);
+		return newsDao.create(testNews);
 	}
 
 	@SmallTest
 	public void testReadNewsUseId() {
 		long id = createNews("guid1");
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			NewsDao newsDao = new NewsDao();
-			News n = newsDao.read(db, id);
-			
-			assertEquals(testNews, n);
-			
-			newsDao.delete(db, id);
-		} finally {
-			db.close();
-		}
+		ContentResolver cr = getContext().getContentResolver();
+		NewsDao newsDao = new NewsDao(cr);
+		News n = newsDao.read(id);
+		
+		assertEquals(testNews, n);
+		
+		newsDao.delete(id);
 	}
 	
 	@SmallTest
 	public void testReadNewsUseGuid() {
 		createNews("guid1");
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			NewsDao newsDao = new NewsDao();
-			News n = newsDao.read(db, "guid1");
+		ContentResolver cr = getContext().getContentResolver();	
+		NewsDao newsDao = new NewsDao(cr);
+		News n = newsDao.read("guid1");
 
-			assertEquals(testNews, n);
-			
-		} finally {
-			db.close();
-		}
+		assertEquals(testNews, n);
+		newsDao.delete("guid1");
 	}
 	
 	@SmallTest
@@ -89,22 +88,18 @@ public class NewsDaoTest extends AndroidTestCase {
 		long linkId1 = it.next().getId();
 		long linkId2 = it.next().getId();
 		long linkId3 = it.next().getId();
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			NewsLinksDao nlinksDao = new NewsLinksDao();
-			NewsDao newsDao = new NewsDao();
-			newsDao.delete(db, "guid1");
-			News n = newsDao.read(db, "guid1");
-			assertNull(n);
-			Link l = nlinksDao.read(db, linkId1);
-			assertNull(l);
-			l = nlinksDao.read(db, linkId2);
-			assertNull(l);
-			l = nlinksDao.read(db, linkId3);
-			assertNull(l);
-		} finally {
-			db.close();
-		}
+		ContentResolver cr = getContext().getContentResolver();
+		NewsLinksDao nlinksDao = new NewsLinksDao(cr);
+		NewsDao newsDao = new NewsDao(cr);
+		newsDao.delete("guid1");
+		News n = newsDao.read("guid1");
+		assertNull(n);
+		Link l = nlinksDao.read(linkId1);
+		assertNull(l);
+		l = nlinksDao.read(linkId2);
+		assertNull(l);
+		l = nlinksDao.read(linkId3);
+		assertNull(l);
 	}
 	
 	@SmallTest
@@ -114,74 +109,66 @@ public class NewsDaoTest extends AndroidTestCase {
 		long linkId1 = it.next().getId();
 		long linkId2 = it.next().getId();
 		long linkId3 = it.next().getId();
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			NewsLinksDao nlinksDao = new NewsLinksDao();
-			NewsDao newsDao = new NewsDao();
-			newsDao.delete(db, id);
-			News n = newsDao.read(db, id);
-			assertNull(n);
-			Link l = nlinksDao.read(db, linkId1);
-			assertNull(l);
-			l = nlinksDao.read(db, linkId2);
-			assertNull(l);
-			l = nlinksDao.read(db, linkId3);
-			assertNull(l);
-		} finally {
-			db.close();
-		}
+		ContentResolver cr = getContext().getContentResolver();
+		NewsLinksDao nlinksDao = new NewsLinksDao(cr);
+		NewsDao newsDao = new NewsDao(cr);
+		newsDao.delete(id);
+		News n = newsDao.read(id);
+		assertNull(n);
+		Link l = nlinksDao.read(linkId1);
+		assertNull(l);
+		l = nlinksDao.read(linkId2);
+		assertNull(l);
+		l = nlinksDao.read(linkId3);
+		assertNull(l);
 	}
 	
 	@SmallTest
 	public void testUpdateNews() {
 		long id = createNews("guid1");
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			NewsDao newsDao = new NewsDao();
-			News n = newsDao.read(db, "guid1");
-			assertEquals("News 1", n.getTitle());
-	
-			n.setBriefText("newbrief1");
-			n.setFullText("newfull1");
-			n.setGuid("newguid1");
-			n.setImageCaption("newcaption1");
-			n.setImageCredits("newcredits1");
-			n.setImageLink("newimagelink1");
-			n.setLink("newlink1");
-			n.setPubDate(date);
-			n.setRubric(Rubrics.ECONOMICS_FINANCE);
-			n.setRubricUpdateNeed(false);
-			n.setTitle("newtitle1");
-			
-			Iterator<Link> it = n.getLinks().iterator();
-			Date dd = new Date();
-			Link l = it.next();
-			l.setDate(dd);
-			l.setTitle("new title");
-			l.setUrl("http://www.mail.ru");
-			
-			l = it.next();
-			l.setDate(dd);
-			l.setTitle("new title 1");
-			l.setUrl("http://www.gmail.com");
-			
-			l = it.next();
-			l.setDate(dd);
-			l.setTitle("new title 2");
-			l.setUrl("http://www.htc.com");
-			
-			newsDao.update(db, n);
-			
-			News n1 = newsDao.read(db, id);
+		ContentResolver cr = getContext().getContentResolver();	
+		NewsDao newsDao = new NewsDao(cr);
+		News n = newsDao.read("guid1");
+		assertEquals("News 1", n.getTitle());
 
-			assertEquals(n, n1);
-			
-			newsDao.delete(db, "newguid1");
-			n = newsDao.read(db, "newguid1");
-			assertNull(n);
-		} finally {
-			db.close();
-		}
+		n.setBriefText("newbrief1");
+		n.setFullText("newfull1");
+		n.setGuid("newguid1");
+		n.setImageCaption("newcaption1");
+		n.setImageCredits("newcredits1");
+		n.setImageLink("newimagelink1");
+		n.setLink("newlink1");
+		n.setPubDate(date);
+		n.setRubric(Rubrics.ECONOMICS_FINANCE);
+		n.setRubricUpdateNeed(false);
+		n.setTitle("newtitle1");
+		
+		Iterator<Link> it = n.getLinks().iterator();
+		Date dd = new Date();
+		Link l = it.next();
+		l.setDate(dd);
+		l.setTitle("new title");
+		l.setUrl("http://www.mail.ru");
+		
+		l = it.next();
+		l.setDate(dd);
+		l.setTitle("new title 1");
+		l.setUrl("http://www.gmail.com");
+		
+		l = it.next();
+		l.setDate(dd);
+		l.setTitle("new title 2");
+		l.setUrl("http://www.htc.com");
+		
+		newsDao.update(n);
+		
+		News n1 = newsDao.read(id);
+
+		assertEquals(n, n1);
+		
+		newsDao.delete("newguid1");
+		n = newsDao.read("newguid1");
+		assertNull(n);
 	}
 	
 	@SmallTest
@@ -197,29 +184,19 @@ public class NewsDaoTest extends AndroidTestCase {
 			createNews(id);
 		}
 		
-		NewsDao newsDao = new NewsDao();
-		SQLiteDatabase db = dbHelper.getReadableDatabase();
-		try {
-			Collection<String> allIds = newsDao.readAllKeys(db);
-			List<String> sortedIds = new ArrayList<String>(allIds);
-			Collections.sort(sortedIds);
-			
-			assertEquals(ids, sortedIds);
-			
-			for (String id : ids) {
-				newsDao.delete(db, id);
-			}
-			
-			allIds = newsDao.readAllKeys(db);
-			assertEquals(allIds.size(), 0);
-		} finally {
-			db.close();
+		ContentResolver cr = getContext().getContentResolver();	
+		NewsDao newsDao = new NewsDao(cr);
+		Collection<String> allIds = newsDao.readAllKeys();
+		List<String> sortedIds = new ArrayList<String>(allIds);
+		Collections.sort(sortedIds);
+		
+		assertEquals(ids, sortedIds);
+		
+		for (String id : ids) {
+			newsDao.delete(id);
 		}
-	}
-	
-	@Override
-	protected void tearDown() throws Exception {
-		super.tearDown();
-		context.deleteDatabase(LentaDbHelper.DATABASE_NAME);
+		
+		allIds = newsDao.readAllKeys();
+		assertEquals(allIds.size(), 0);
 	}
 }
